requirements.txt
configparser==4.0.2
psycopg2-binary==2.8.4
